requirements.txt
alabaster==0.7.10
Babel==2.5.1
certifi==2019.3.9
chardet==3.0.4
django-filter==1.0.4
Django==1.11.20
djangorestframework==3.9.4
django-cors-headers==2.5.2
docutils==0.14
dramatiq==1.4.0
idna==2.8
imagesize==0.7.1
Jinja2==2.10.1
MarkupSafe==1.0
psycopg2==2.7.3.1
Pygments==2.2.0
pytz==2018.9
requests==2.21.0
six==1.11.0
snowballstemmer==1.2.1
sphinx-rtd-theme==0.2.4
Sphinx==1.6.4
sphinxcontrib-websupport==1.0.1
urllib3==1.24.3
